Mention getting long-term contributors as primary goal of GSoC participation
This will probably make the Google folks happy, as this is officially the main
purpose of the GSoC program...

More importantly, past experience shows that even very skilled students are
seldom able to leave a project in such a state that it is really useful for us
without any followup work -- treating GSoC as a way to get code written by
means of external sponsoring doesn't really work out. Finding long-term
contributors is *much* more useful in the end.