Rework introduction section in GSoC project ideas list
Make it clear that we explicitely require the students to contact us,
and to submit a patch as part of the application process.

Most of this information is redundant with the introduction section of
the student application form. This is not a problem: If they read it
twice, there is a better chance that they will take it seriously...