gsoc/ideas/driver_glue_code: Update and add back
As zhengda seems MIA, this task is up for takers again.

Update the text to reflect the fact that zhengda already did a lot of
the DDE work. (Actually, the text is almost entirely rewritten...)

Probably this should be split into several distinct tasks for the
various aspects that need work.