dde/guide: Drop extra note about possibly necessary Makefile changes
Even if it should turn out this is not completely obsolete (which I'm
pretty sure it is), the specific cases are already covered in the
instructions; and if any new should pop up, the guide needs to be
updated accordingly. Making the user guess just doesn't fit with such a
step-by-step guide.