dde/guide: Use tabs instead of quote marks
This way we lose the indentation -- but special characters are escaped
automatically. Also, the commands appear in monospace font (as <pre>
blocks) this way, which I think is better.

There might be some prettier option -- but I'm too lazy to check :-) The
formatting of the whole thing is not exactly good anyways...