mach/building: Rearrange bit about installing header files
Instead of treating all the build preparations as part of the header
install, make them part of the general build instructions; and only
separately mention the specific bit about actually installing the
headers.

This should make the overall build process easier to follow; and also
makes it clearer which bit can be left out if installing the headers is
not necessary.