news/2011-q2-ps: s#GNU/Hurd#GNU Hurd# in some section headings
Don't use "GNU/Hurd" when talking about things not realated to
distributions -- as explained in this very same news article... :-)

BTW, "The GNU/Hurd" is *never* correct -- just like nobody would say
"the GNU/Linux", or "the GNU". Only "the Hurd" is special in this
regard. (And sometimes "the GNU Hurd" is also OK, but it only fits in
some contexts! Often just "GNU Hurd" is more appropriate as well.)